backend/requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
Flask-MySQLdb==2.0.0
Flask-Bcrypt==1.0.1
mysqlclient==2.2.8
joblib==1.5.3
numpy==2.4.2
requests==2.32.5
python-dotenv==1.2.2
scikit-learn
